models/003-NaViT/requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# 数据处理和科学计算
numpy>=1.21.0
Pillow>=8.3.0
scipy>=1.7.0

# 机器学习指标和工具
torchmetrics>=0.11.0
scikit-learn>=1.0.0

# 可视化和日志
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0

# 数据加载和处理
pandas>=1.3.0
h5py>=3.6.0

# 进度条和实用工具
tqdm>=4.62.0
colorama>=0.4.4

# 配置文件处理
PyYAML>=6.0

# 图像处理增强
albumentations>=1.3.0

# 模型优化和量化（可选）
timm>=0.6.0

# 开发和调试工具
ipython>=7.0.0
jupyter>=1.0.0

# 类型检查（可选）
mypy>=0.910

# 代码格式化（可选）
black>=22.0.0
flake8>=4.0.0